refactor: move getDeviceControlsState to matter-device-binding-card.
