+++ src/panels/config/integrations/integration-panels/matter/dialog-matter-node-binding.ts
@@ -1,4 +1,3 @@
-import type { HassEntity } from "home-assistant-js-websocket";
 import type { CSSResultGroup, PropertyValues } from "lit";
 import { css, html, LitElement, nothing } from "lit";
 import { customElement, property, state } from "lit/decorators";
@@ -15,12 +14,12 @@ import type { DeviceRegistryEntry } from "../../../../../data/device_registry";
 import type { MatterNodeBindingDialogParams } from "./show-dialog-matter-node-binding";
 import type { MatterDeviceMapper } from "./matter-binding-node-device-mapper";
 import type { MatterNodeBinding } from "../../../../../data/matter";
-import type { EntityRegistryDisplayEntry } from "../../../../../data/entity_registry";
 
 import { fireEvent } from "../../../../../common/dom/fire_event";
 import { stopPropagation } from "../../../../../common/dom/stop_propagation";
 import { haStyle, haStyleDialog } from "../../../../../resources/styles";
 import { createCloseHeading } from "../../../../../components/ha-dialog";
+import { getDeviceControlsState } from "./matter-device-binding-card";
 
 export interface ItemSelectedEvent {
   target?: HaSelect;
@@ -88,34 +87,6 @@ class DialogMatterNodeBinding extends LitElement {
     }
   }
 
-  getDeviceControlsState(
-    hass: HomeAssistant,
-    device: DeviceRegistryEntry
-  ): HassEntity | undefined {
-    // Helper function to find the first matching entity
-    const findEntity = (
-      predicate: (entity: EntityRegistryDisplayEntry) => boolean
-    ): HassEntity | undefined => {
-      const entity = Object.values(hass.entities).find(predicate);
-      return entity ? hass.states[entity.entity_id] : undefined;
-    };
-
-    // Try to find a control entity (no category)
-    const controlState = findEntity(
-      (entity: EntityRegistryDisplayEntry) =>
-        entity.device_id === device.id && entity.entity_category === undefined
-    );
-    if (controlState) return controlState;
-
-    // Fallback to config "Identify" entity
-    return findEntity(
-      (entity: EntityRegistryDisplayEntry) =>
-        entity.device_id === device.id &&
-        entity.entity_category === "config" &&
-        entity.name === "Identify"
-    );
-  }
-
   protected updated(changedProperties: PropertyValues): void {
     if (changedProperties.has("hass")) {
       this._bindableDevices = Object.values(this.hass.devices).filter(
@@ -150,7 +121,7 @@ class DialogMatterNodeBinding extends LitElement {
                   <ha-state-icon
                     slot="graphic"
                     .hass=${this.hass}
-                    .stateObj=${this.getDeviceControlsState(this.hass, device)}
+                    .stateObj=${getDeviceControlsState(this.hass, device)}
                   ></ha-state-icon>
                 </ha-list-item>
               `
